[AN] refactor: HearitMapper 생성 #111
