+++ android/app/src/main/java/com/onair/hearit/data/dto/RandomHearitResponse.kt
@@ -1,6 +1,5 @@
 package com.onair.hearit.data.dto
 
-import com.onair.hearit.domain.RandomHearitItem
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -19,13 +18,3 @@ data class RandomHearitResponse(
     @SerialName("title")
     val title: String,
 )
-
-fun RandomHearitResponse.toDomain(): RandomHearitItem =
-    RandomHearitItem(
-        id = this.id,
-        title = this.title,
-        summary = this.summary,
-        source = this.source,
-        playTime = this.playTime,
-        createdAt = this.createdAt,
-    )
